# Conflicts:
#	GGJ-2023-NATDI/Assets/Scripts/MushroomControls.cs

+++ GGJ-2023-NATDI/Assets/Scripts/SpawnerService.cs
@@ -32,6 +32,11 @@ public class SpawnerService : Service, IInject
 
     public void DespawnMushroom(Mushroom mushroom)
     {
+        if (mushroom == null)
+        {
+           Debug.LogError("mushroom already destroyed");
+           return;
+        }
         _collectionService.RemoveMushroom(mushroom);
         Destroy(mushroom.gameObject);
     }
